@@ -1,16 +1,18 @@
+// ignore_for_file: avoid_print
+
 import 'package:flutter/material.dart';
 import 'package:flutter_unity_widget/flutter_unity_widget.dart';
 import 'package:pointer_interceptor/pointer_interceptor.dart';
 
 class ApiScreen extends StatefulWidget {
-  ApiScreen({Key key}) : super(key: key);
+  const ApiScreen({Key? key}) : super(key: key);
 
   @override
-  _ApiScreenState createState() => _ApiScreenState();
+  State<ApiScreen> createState() => _ApiScreenState();
 }
 
 class _ApiScreenState extends State<ApiScreen> {
-  UnityWidgetController _unityWidgetController;
+  UnityWidgetController? _unityWidgetController;
   double _sliderValue = 0.0;
 
   @override
@@ -20,7 +22,7 @@ class _ApiScreenState extends State<ApiScreen> {
 
   @override
   void dispose() {
-    _unityWidgetController.dispose();
+    _unityWidgetController?.dispose();
     super.dispose();
   }
 
@@ -28,7 +30,7 @@ class _ApiScreenState extends State<ApiScreen> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('API Screen'),
+        title: const Text('API Screen'),
       ),
       body: Card(
         margin: const EdgeInsets.all(8),
@@ -38,14 +40,12 @@ class _ApiScreenState extends State<ApiScreen> {
         ),
         child: Stack(
           children: [
-            Container(
-              child: UnityWidget(
-                onUnityCreated: onUnityCreated,
-                onUnityMessage: onUnityMessage,
-                onUnitySceneLoaded: onUnitySceneLoaded,
-                fullscreen: false,
-                useAndroidViewSurface: false,
-              ),
+            UnityWidget(
+              onUnityCreated: onUnityCreated,
+              onUnityMessage: onUnityMessage,
+              onUnitySceneLoaded: onUnitySceneLoaded,
+              fullscreen: false,
+              useAndroidViewSurface: false,
             ),
             PointerInterceptor(
               child: Positioned(
@@ -57,8 +57,8 @@ class _ApiScreenState extends State<ApiScreen> {
                   child: Column(
                     mainAxisSize: MainAxisSize.min,
                     children: <Widget>[
-                      Padding(
-                        padding: const EdgeInsets.only(top: 20),
+                      const Padding(
+                        padding: EdgeInsets.only(top: 20),
                         child: Text("Rotation speed:"),
                       ),
                       Slider(
@@ -78,27 +78,27 @@ class _ApiScreenState extends State<ApiScreen> {
                           children: [
                             MaterialButton(
                               onPressed: () {
-                                _unityWidgetController.quit();
+                                _unityWidgetController?.quit();
                               },
-                              child: Text("Quit"),
+                              child: const Text("Quit"),
                             ),
                             MaterialButton(
                               onPressed: () {
-                                _unityWidgetController.create();
+                                _unityWidgetController?.create();
                               },
-                              child: Text("Create"),
+                              child: const Text("Create"),
                             ),
                             MaterialButton(
                               onPressed: () {
-                                _unityWidgetController.pause();
+                                _unityWidgetController?.pause();
                               },
-                              child: Text("Pause"),
+                              child: const Text("Pause"),
                             ),
                             MaterialButton(
                               onPressed: () {
-                                _unityWidgetController.resume();
+                                _unityWidgetController?.resume();
                               },
-                              child: Text("Resume"),
+                              child: const Text("Resume"),
                             ),
                           ],
                         ),
@@ -110,21 +110,21 @@ class _ApiScreenState extends State<ApiScreen> {
                             MaterialButton(
                               onPressed: () async {
                                 await _unityWidgetController
-                                    .openInNativeProcess();
+                                    ?.openInNativeProcess();
                               },
-                              child: Text("Open Native"),
+                              child: const Text("Open Native"),
                             ),
                             MaterialButton(
                               onPressed: () {
-                                _unityWidgetController.unload();
+                                _unityWidgetController?.unload();
                               },
-                              child: Text("Unload"),
+                              child: const Text("Unload"),
                             ),
                             MaterialButton(
                               onPressed: () {
-                                _unityWidgetController.quit();
+                                _unityWidgetController?.quit();
                               },
-                              child: Text("Silent Quit"),
+                              child: const Text("Silent Quit"),
                             ),
                           ],
                         ),
@@ -141,7 +141,7 @@ class _ApiScreenState extends State<ApiScreen> {
   }
 
   void setRotationSpeed(String speed) {
-    _unityWidgetController.postMessage(
+    _unityWidgetController?.postMessage(
       'Cube',
       'SetRotationSpeed',
       speed,
@@ -152,13 +152,17 @@ class _ApiScreenState extends State<ApiScreen> {
     print('Received message from unity: ${message.toString()}');
   }
 
-  void onUnitySceneLoaded(SceneLoaded scene) {
-    print('Received scene loaded from unity: ${scene.name}');
-    print('Received scene loaded from unity buildIndex: ${scene.buildIndex}');
+  void onUnitySceneLoaded(SceneLoaded? scene) {
+    if (scene != null) {
+      print('Received scene loaded from unity: ${scene.name}');
+      print('Received scene loaded from unity buildIndex: ${scene.buildIndex}');
+    } else {
+      print('Received scene loaded from unity: null');
+    }
   }
 
   // Callback that connects the created controller to the unity controller
   void onUnityCreated(controller) {
-    this._unityWidgetController = controller;
+    _unityWidgetController = controller;
   }
 }
