@@ -3,17 +3,17 @@ import 'package:flutter_unity_widget/flutter_unity_widget.dart';
 import 'package:pointer_interceptor/pointer_interceptor.dart';
 
 class NoInteractionScreen extends StatefulWidget {
-  NoInteractionScreen({Key key}) : super(key: key);
+  const NoInteractionScreen({Key? key}) : super(key: key);
 
   @override
-  _NoInteractionScreenState createState() => _NoInteractionScreenState();
+  State<NoInteractionScreen> createState() => _NoInteractionScreenState();
 }
 
 class _NoInteractionScreenState extends State<NoInteractionScreen> {
   static final GlobalKey<ScaffoldState> _scaffoldKey =
       GlobalKey<ScaffoldState>();
 
-  UnityWidgetController _unityWidgetController;
+  UnityWidgetController? _unityWidgetController;
 
   @override
   void initState() {
@@ -22,7 +22,7 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
 
   @override
   void dispose() {
-    _unityWidgetController.dispose();
+    _unityWidgetController?.dispose();
     super.dispose();
   }
 
@@ -31,7 +31,7 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
     return Scaffold(
       key: _scaffoldKey,
       appBar: AppBar(
-        title: Text('No Interaction Screen'),
+        title: const Text('No Interaction Screen'),
       ),
       body: Card(
         margin: const EdgeInsets.all(8),
@@ -46,7 +46,7 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
               onUnityMessage: onUnityMessage,
               onUnitySceneLoaded: onUnitySceneLoaded,
               useAndroidViewSurface: true,
-              borderRadius: BorderRadius.all(Radius.circular(70)),
+              borderRadius: const BorderRadius.all(Radius.circular(70)),
             ),
             PointerInterceptor(
               child: Positioned(
@@ -57,7 +57,7 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
                   onPressed: () {
                     Navigator.of(context).pushNamed('/simple');
                   },
-                  child: Text('Switch Flutter Screen'),
+                  child: const Text('Switch Flutter Screen'),
                 ),
               ),
             ),
@@ -68,7 +68,7 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
   }
 
   void setRotationSpeed(String speed) {
-    _unityWidgetController.postMessage(
+    _unityWidgetController?.postMessage(
       'Cube',
       'SetRotationSpeed',
       speed,
@@ -79,14 +79,18 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
     print('Received message from unity: ${message.toString()}');
   }
 
-  void onUnitySceneLoaded(SceneLoaded scene) {
-    print('Received scene loaded from unity: ${scene.name}');
-    print('Received scene loaded from unity buildIndex: ${scene.buildIndex}');
+  void onUnitySceneLoaded(SceneLoaded? scene) {
+    if (scene != null) {
+      print('Received scene loaded from unity: ${scene.name}');
+      print('Received scene loaded from unity buildIndex: ${scene.buildIndex}');
+    } else {
+      print('Received scene loaded from unity: null');
+    }
   }
 
   // Callback that connects the created controller to the unity controller
   void _onUnityCreated(controller) {
     controller.resume();
-    this._unityWidgetController = controller;
+    _unityWidgetController = controller;
   }
 }
