@@ -24,6 +24,7 @@ using System;
 
 using System.Collections.Generic;
 using System.IO;
+using System.Text.RegularExpressions;
 
 using UnityEditor;
 using UnityEditor.iOS.Xcode;
@@ -75,7 +76,7 @@ public static class XcodePostBuild
         pbx.SetBuildProperty(targetGuid, "SKIP_INSTALL", "YES");
 
         // Set some linker flags
-        pbx.SetBuildProperty(projGuid, "ENABLE_BITCODE", "YES");
+        pbx.SetBuildProperty(projGuid, "ENABLE_BITCODE", "NO");
 
         // Persist changes
         pbx.WriteToFile(pbxPath);
@@ -323,7 +324,9 @@ public static class XcodePostBuild
             inScope |= line.Contains("- (void)startUnity:");
             markerDetected |= inScope && line.Contains(TouchedMarker);
 
-            if (inScope && line.Trim() == "}")
+            //Find the end of the startUnity function, a } without any indentation.  (regex: starts with } followed by any whitespace)
+            //Avoid indentation before } as newer unity versions include an if-statement inside this function.
+            if (inScope && Regex.Match(line, @"^}(\s)*$").Success)
             {
                 inScope = false;
 
@@ -333,6 +336,7 @@ public static class XcodePostBuild
                 }
                 else
                 {
+                    //Add a UnityReady notification at the end of the startUnity function.
                     return new string[]
                     {
                         "    // Modified by " + TouchedMarker,
